refactor: respond with JSON

+++ app/user/src/admin.ts
@@ -3,7 +3,40 @@ import { users } from './db/schema';
 import { db } from './utils/db';
 import { eq, ilike } from 'drizzle-orm';
 
+export class FoilCTF_Error extends Error {
+	public statusCode: number;
+	constructor(message: string, statusCode: number) {
+		super(message);
+		this.statusCode = statusCode;
 
+		this.name = 'FoilCTF_Error';
+	}
+
+	toJSON() {
+		return {
+			error: true,
+			message: this.message,
+			status: this.statusCode
+		}
+	}
+}
+
+export class FoilCTF_Success {
+	public statusCode: number;
+	public message: string;
+
+	constructor(message: string, statusCode: number) {
+		this.statusCode = statusCode;
+		this.message = message
+	}
+
+	toJSON() {
+		return {
+			message: this.message,
+			status: this.statusCode
+		}
+	}
+}
 
 export async function listUsers(req: Request, res: Response, next: NextFunction) {
     const limit = Math.max(Number(req.query.limit) || 10, 1);
@@ -12,7 +45,7 @@ export async function listUsers(req: Request, res: Response, next: NextFunction)
 
     const decodedUser = res.locals.user;
     if (decodedUser.role !== "admin")
-        return res.status(403).json();
+        return res.status(403).json(new FoilCTF_Error("Forbidden", 403));
 
     const dbUsers = await db
         .select()
@@ -32,13 +65,21 @@ export async function updateUserRole(req: Request, res: Response, next: NextFunc
     const target = req.params.username as string;
     const newRole = req.body.newRole;
 
+	const decodedUser = res.locals.user;
+    if (decodedUser.role !== "admin")
+        return res.status(403).json(new FoilCTF_Error("Forbidden", 403));
+
     await db.update(users).set({ role: newRole }).where(eq(users.username, target));
-    return res.status(200).json();
+    return res.status(200).json(new FoilCTF_Success("OK", 200));
 }
 
 export async function deleteUser(req: Request, res: Response) {
     const target = req.params.username as string;
 
+	const decodedUser = res.locals.user;
+    if (decodedUser.role !== "admin")
+        return res.status(403).json(new FoilCTF_Error("Forbidden", 403));
+
     await db.delete(users).where(eq(users.username, target));
-    return res.status(204).send();
+    return res.status(204).json(new FoilCTF_Success("No Content", 204));
 }
